pvattest: Create, do, and verify Attestation Measurements
pvattest is a tool to attest an IBM Secure Execution guest.

In a trusted environment, one can create request using
`pvattest create`. To get a measurement of an untrusted
IBM Secure Execution guest call 'pvattest measure'.
Again in a trusted environment, call 'pvattest verify'
to verify that the measurement is the expected one.

The tool runs on s390 and x86.
It has the same requirements like libpv.
Additionally, to measure, the linux kernel must provide
the Ultravisor-userspace-interface  `uvdevice` and be executed
on a z16+ IBM Secure Execution guest.
To generate man pages txt2man is required.

Signed-off-by: Steffen Eiden <seiden@linux.ibm.com>